afternoon update

+++ CADCOP/src/Messages/MsgAlgorithmFactor.java
@@ -6,7 +6,7 @@ public class MsgAlgorithmFactor extends MsgAlgorithm{
     private NodeId recieverNode, senderNode;
 	
     public MsgAlgorithmFactor(NodeId senderNode, NodeId recieverNode, Object context, 
-    		int timeStamp, double delay,double timeCreated ) {
+    		double timeStamp, double delay,double timeCreated ) {
 		super(-1, -1, context, timeStamp, delay, timeCreated);
 		this.recieverNode = recieverNode;
 		this.senderNode = senderNode;
